(#857) Load Chocolatey executable assembly
This commit makes changes to the assembly loader to
load the installed Chocolatey executable as an
assembly instead of the statically referenced
DLL file

These changes are necessary to ensure that there will
be minimal incompatibilities between running Chocolatey CLI
and Chocolatey GUI for both FOSS and Licensed users.
