+++ Source/ChocolateyGui/App.xaml.cs
@@ -6,13 +6,13 @@
 // --------------------------------------------------------------------------------------------------------------------
 
 using System;
+using System.IO;
 using System.Reflection;
 using System.Windows;
 using Autofac;
-using chocolatey;
-using chocolatey.infrastructure.registration;
 using ChocolateyGui.Common.Enums;
 using ChocolateyGui.Common.Services;
+using ChocolateyGui.Common.Startup;
 using ChocolateyGui.Common.Windows;
 using ChocolateyGui.Common.Windows.Theming;
 
@@ -32,30 +32,49 @@ namespace ChocolateyGui
             {
                 var requestedAssembly = new AssemblyName(args.Name);
 
-#if FORCE_CHOCOLATEY_OFFICIAL_KEY
-                var chocolateyGuiPublicKey = Bootstrapper.OfficialChocolateyPublicKey;
-#else
-                var chocolateyGuiPublicKey = Bootstrapper.UnofficialChocolateyPublicKey;
-#endif
-
                 try
                 {
-                    if (requestedAssembly.get_public_key_token().is_equal_to(chocolateyGuiPublicKey)
-                        && requestedAssembly.Name.is_equal_to(Bootstrapper.ChocolateyGuiCommonAssemblySimpleName))
+                    if (string.Equals(requestedAssembly.Name, "chocolatey", StringComparison.OrdinalIgnoreCase))
                     {
-                        return AssemblyResolution.resolve_or_load_assembly(
-                            Bootstrapper.ChocolateyGuiCommonAssemblySimpleName,
-                            requestedAssembly.get_public_key_token(),
-                            Bootstrapper.ChocolateyGuiCommonAssemblyLocation).UnderlyingType;
+                        var installDir = Environment.GetEnvironmentVariable("ChocolateyInstall");
+                        if (string.IsNullOrEmpty(installDir))
+                        {
+                            var rootDrive = Path.GetPathRoot(Assembly.GetExecutingAssembly().Location);
+                            if (string.IsNullOrEmpty(rootDrive))
+                            {
+                                return null; // TODO: Maybe return the chocolatey.dll file instead?
+                            }
+
+                            installDir = Path.Combine(rootDrive, "ProgramData", "chocolatey");
+                        }
+
+                        var assemblyLocation = Path.Combine(installDir, "choco.exe");
+
+                        return AssemblyResolver.ResolveOrLoadAssembly("choco", string.Empty, assemblyLocation);
                     }
 
-                    if (requestedAssembly.get_public_key_token().is_equal_to(chocolateyGuiPublicKey)
-                        && requestedAssembly.Name.is_equal_to(Bootstrapper.ChocolateyGuiCommonWindowsAssemblySimpleName))
+#if FORCE_CHOCOLATEY_OFFICIAL_KEY
+                    var chocolateyGuiPublicKey = Bootstrapper.OfficialChocolateyPublicKey;
+#else
+                    var chocolateyGuiPublicKey = Bootstrapper.UnofficialChocolateyPublicKey;
+#endif
+
+                    if (AssemblyResolver.DoesPublicKeyTokenMatch(requestedAssembly, chocolateyGuiPublicKey)
+                        && string.Equals(requestedAssembly.Name, Bootstrapper.ChocolateyGuiCommonAssemblySimpleName, StringComparison.OrdinalIgnoreCase))
                     {
-                        return AssemblyResolution.resolve_or_load_assembly(
+                        return AssemblyResolver.ResolveOrLoadAssembly(
+                            Bootstrapper.ChocolateyGuiCommonAssemblySimpleName,
+                            AssemblyResolver.GetPublicKeyToken(requestedAssembly),
+                            Bootstrapper.ChocolateyGuiCommonAssemblyLocation);
+                    }
+
+                    if (AssemblyResolver.DoesPublicKeyTokenMatch(requestedAssembly, chocolateyGuiPublicKey)
+                        && string.Equals(requestedAssembly.Name, Bootstrapper.ChocolateyGuiCommonWindowsAssemblySimpleName, StringComparison.OrdinalIgnoreCase))
+                    {
+                        return AssemblyResolver.ResolveOrLoadAssembly(
                             Bootstrapper.ChocolateyGuiCommonWindowsAssemblySimpleName,
-                            requestedAssembly.get_public_key_token(),
-                            Bootstrapper.ChocolateyGuiCommonWindowsAssemblyLocation).UnderlyingType;
+                            AssemblyResolver.GetPublicKeyToken(requestedAssembly),
+                            Bootstrapper.ChocolateyGuiCommonWindowsAssemblyLocation);
                     }
                 }
                 catch (Exception ex)
